i3/i3init/pyproject.toml
[project]
name = "i3init"
description = ''
readme = "README.md"
requires-python = ">=3.10"
keywords = []
authors = [
  { name = "Reinis Taukulis", email = "reinis.taukulis@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "i3ipc",
  "python-rofi-phfn",
]
version = "0.0.3-dev"

[license]
file = "LICENSE.txt"

[project.scripts]
i3init = "i3init.cli:main_cli"

# [project.urls]
# Documentation = "https://github.com/unknown/i3init#readme"
# Issues = "https://github.com/unknown/i3init/issues"
# Source = "https://github.com/unknown/i3init"

[tool.uv]
package = true

[tool.pyright]
include = ["src", "tests"]
stubPath = "typings"
reportMissingImports = true
reportMissingTypeStubs = false
